+++ shared/src/simulation/behaviors/answer-requests.ts
@@ -125,7 +125,6 @@ export const answerRequestsBehavior: SimulationBehavior<AnswerRequestsBehaviorSt
                                 DelayEventActivityState.create(
                                     nextUUID(draftState),
                                     ResourceRequiredEvent.create(
-                                        '',
                                         simulatedRegion.id,
                                         VehicleResource.create(
                                             vehiclesNotAvailable