+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project does **not** adhere to [Semantic Versioning](https://semver.org
 
 -   Add behaviors button now opens towards the top.
 
+### Fixed
+
+-   Removed the id property of events.
+
 ## [0.5.1] - 2023-05-09
 
 ### Fixed